+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]## [0.1.2.2b2]
 ### Changed
 - DoDashApp.read_scenario_tables_from_db_cached only reads tables in the schema, skips others without warning.
+- Enabled virtualization in DashTables (Prepare data and Explore Solution)
 
 ## [0.1.2.2b1] - 2022-10-17
 ### Changed